server/requirements.txt
fastapi
pydantic
pymongo
uvicorn[standard]
python-dotenv
